Suite for workers. Separate common node-worker tests into isomorphic
