+++ spec/javascripts/integration/index.worker.js
@@ -1,0 +1,2 @@
+var sharedTestsContext = require.context("./core", true, /_spec$/);
+sharedTestsContext.keys().forEach(sharedTestsContext);
