+++ spec/javascripts/unit/index.worker.js
@@ -4,5 +4,5 @@ sharedTestsContext.keys().forEach(sharedTestsContext);
 var nodeTestsContext = require.context("./isomorphic", true, /_spec$/);
 nodeTestsContext.keys().forEach(nodeTestsContext);
 
-var nodeTestsContext = require.context("./node", true, /_spec$/);
-nodeTestsContext.keys().forEach(nodeTestsContext)
+var nodeTestsContext = require.context("./worker", true, /_spec$/);
+nodeTestsContext.keys().forEach(nodeTestsContext);
